Skill ut tabellrad-logikk i egen komponent
